docs(readme): add missing information to readme
The readme was missing some required instructions to properly get set up, such as the installation of the karma cli. This commit adds those instructions. It also includes instructions explaining how to run the samples. Some general cleanup and organization to the file was done as part of the process of adding these additional details.

Closes #22